Refactor boundary-mode architecture: BMO as SpaceOperator analog, thin ModeEigenSolver
Rework the 2D boundary mode class structure so responsibilities match
palace's existing operator / solver / driver layering, and so WavePort
and BoundaryMode share the same ModeEigenSolver and assembly primitives.

BoundaryModeOperator is now the SpaceOperator analog: it owns the FE
hierarchies and boundary operators, assembles the frequency-independent
block matrices (Atn, Btn = -Atn^T, Btt) in its ctor, and exposes
AssembleAtt(omega, sigma) / AssembleAnn(omega) / ApplyVDBackTransform /
ComputePoyntingPower for the eigenvalue solver. It takes MaterialOperator
by const ref and an optional 3D normal for tangent-plane projection, and
no longer knows anything about submesh extraction.

The shared block-assembly primitives live in palace/models/mode_assembly.
{hpp,cpp} as free functions parameterized by FE space + material +
boundary operators + normal — used by BoundaryModeOperator on the 2D
domain path and by ModeEigenSolver's bare ctor on the 3D wave port
submesh path.

ModeEigenSolver is now the thin eigenvalue solver in palace/models/
modeeigensolver.{hpp,cpp}: builds the shifted block system from an
operator (or bare FE spaces), drives KSP + SLEPc/ARPACK, returns
eigenpairs and propagation constants. Two ctors — bare FE-space for
WavePort, BoundaryModeOperator& for BoundaryMode — replace the previous
ModeEigenSolverMultigridConfig plumbing that BoundaryModeOperator was
the only producer of (wave-port BCs always passed nullptr).

mesh::ExtractStandalone2DSubmesh and the ProjectSubmeshTo2D(ParMesh&)
overload go away; submesh extraction is now composed from the templated
SubMesh primitives (CreateFromBoundary / RemapSubMeshAttributes /
RemapSubMeshBdrAttributes / AddSubMeshInternalBoundaryElements /
ProjectSubmeshTo2D) and happens in the driver's serial preprocessing
hook. The composition is BM-specific, so SubmeshFrame and the
ExtractBoundary2DSubmesh helper live in an anonymous namespace in
boundarymodesolver.cpp; the reusable primitives in the mesh layer stay
public for WavePort.

BoundaryModeSolver::Preprocess (the BaseSolver hook override) extracts
the 2D submesh on the pre-partition serial mesh, handles the BM-specific
waveport->PEC relabeling locally, broadcasts the tangent frame to non-
holding ranks, and bakes the frame into iodata so downstream code is
frame-agnostic — RotateMaterialDefinitions rotates material tensors into
the local frame in place, and 3D impedance / voltage path coordinates
project to 2D local coords so PostOperator reads them as native 2D
paths. Lc resolution and nondimensionalization delegate to the base.

Solve constructs MaterialOperator natively (no post-hoc rotation, since
the iodata-side definitions are already in the local frame), constructs
BoundaryModeOperator, and builds a thin ModeEigenSolver via the
BoundaryModeOperator ctor.

PostOperator<BOUNDARYMODE>::MeasureAndPrintAll drops the Poynting power
P parameter and computes it internally via mode_op.ComputePoyntingPower;
the driver keeps one ComputePoyntingPower call for the pre-measurement
power normalization scale but no longer threads P through the signature.
WavePort uses the new bare-ctor ModeEigenSolver and back-transforms via
mode_assembly::ApplyVDBackTransform instead of a member method. Trailing-
underscore member names dropped throughout.

No wave-port BC behaviour change.
