+++ palace/drivers/boundarymodesolver.cpp
@@ -3,21 +3,210 @@
 
 #include "boundarymodesolver.hpp"
 
+#include <algorithm>
+#include <complex>
+#include <unordered_set>
+#include "linalg/errorestimator.hpp"
 #include "linalg/operator.hpp"
 #include "models/boundarymodeoperator.hpp"
+#include "models/materialoperator.hpp"
+#include "models/modeeigensolver.hpp"
 #include "models/postoperator.hpp"
 #include "utils/communication.hpp"
+#include "utils/geodata.hpp"
+#include "utils/iodata.hpp"
 #include "utils/timer.hpp"
 
 namespace palace
 {
 
+namespace
+{
+
+// Tangent frame of the extracted 2D submesh in the parent 3D coordinates. Used by
+// Preprocess to rotate material tensors and project path coordinates into the local
+// frame; does not escape the driver.
+struct SubmeshFrame
+{
+  mfem::Vector centroid, e1, e2, normal;
+};
+
+// Full 3D-boundary → 2D-submesh pipeline on the pre-partitioned serial mesh:
+// CreateFromBoundary → attribute remap → add internal-boundary edges → 3D→2D
+// projection. Runs on ranks that hold the serial mesh; the returned 2D mesh has its
+// own node coordinates and no longer references the parent.
+std::unique_ptr<mfem::Mesh>
+ExtractBoundary2DSubmesh(mfem::Mesh &parent, const mfem::Array<int> &surface_attrs,
+                         const std::vector<int> &internal_bdr_attrs, SubmeshFrame &frame)
+{
+  auto sub = std::make_unique<mfem::SubMesh>(
+      mfem::SubMesh::CreateFromBoundary(parent, surface_attrs));
+  mesh::RemapSubMeshAttributes(*sub);
+  mesh::RemapSubMeshBdrAttributes(*sub, surface_attrs);
+  mesh::AddSubMeshInternalBoundaryElements(*sub, surface_attrs, internal_bdr_attrs);
+  frame.normal = mesh::ProjectSubmeshTo2D(*sub, frame.centroid, frame.e1, frame.e2);
+  return sub;
+}
+
+// Project each 3D point in `path` onto the 2D local frame (centroid, e1, e2). Leaves
+// entries that are already 2D untouched. `path` is the iodata-side vector-of-vectors
+// representation; we write back 2-element entries so the downstream PostOperator reads
+// them as 2D coordinates.
+void ProjectPathTo2D(std::vector<std::vector<double>> &path, const mfem::Vector &centroid,
+                     const mfem::Vector &e1, const mfem::Vector &e2)
+{
+  for (auto &p : path)
+  {
+    if (p.size() != 3)
+    {
+      continue;
+    }
+    mfem::Vector p3d(p.data(), 3);
+    mfem::Vector p2d = mesh::Project3Dto2D(p3d, centroid, e1, e2);
+    p.assign({p2d(0), p2d(1)});
+  }
+}
+
+}  // namespace
+
+BoundaryModeSolver::BoundaryModeSolver(const IoData &iodata, bool root, int size,
+                                       int num_thread, const char *git_tag)
+  : BaseSolver(iodata, root, size, num_thread, git_tag)
+{
+}
+
+void BoundaryModeSolver::Preprocess(IoData &iodata, std::unique_ptr<mfem::Mesh> &smesh,
+                                    MPI_Comm comm) const
+{
+  const auto &bm = iodata.solver.boundary_mode;
+  if (bm.attributes.empty())
+  {
+    // Direct 2D path: caller-supplied mesh is already the solve mesh.
+    BaseSolver::Preprocess(iodata, smesh, comm);
+    return;
+  }
+
+  SubmeshFrame frame;
+  frame.centroid.SetSize(3);
+  frame.centroid = 0.0;
+  frame.e1.SetSize(3);
+  frame.e1 = 0.0;
+  frame.e2.SetSize(3);
+  frame.e2 = 0.0;
+  frame.normal.SetSize(3);
+  frame.normal = 0.0;
+
+  // Ranks holding the serial mesh run the extraction locally via mfem::SubMesh;
+  // non-holders receive the frame via broadcast below.
+  if (smesh)
+  {
+    MFEM_VERIFY(smesh->Dimension() == 3,
+                "BoundaryMode with \"Attributes\" requires a 3D input mesh!");
+    Mpi::Print(" Extracting 2D submesh from 3D boundary attributes...\n");
+
+    mfem::Array<int> attr_list;
+    attr_list.Append(bm.attributes.data(), bm.attributes.size());
+
+    // Parent boundary face types whose intersection with the cross-section should
+    // become 2D boundary edges: BCs that pin the tangential E-field (PEC, auxpec,
+    // impedance, conductivity, farfield) plus other-waveports (relabeled to PEC below).
+    std::vector<int> internal_bdr_attrs;
+    const auto &bdr = iodata.boundaries;
+    auto append = [&](const auto &src)
+    { internal_bdr_attrs.insert(internal_bdr_attrs.end(), src.begin(), src.end()); };
+    append(bdr.pec.attributes);
+    append(bdr.auxpec.attributes);
+    for (const auto &d : bdr.impedance)
+    {
+      append(d.attributes);
+    }
+    for (const auto &d : bdr.conductivity)
+    {
+      append(d.attributes);
+    }
+    append(bdr.farfield.attributes);
+
+    std::unordered_set<int> other_waveport_attrs;
+    for (const auto &[idx, data] : bdr.waveport)
+    {
+      for (auto a : data.attributes)
+      {
+        if (std::find(bm.attributes.begin(), bm.attributes.end(), a) != bm.attributes.end())
+        {
+          continue;
+        }
+        other_waveport_attrs.insert(a);
+        internal_bdr_attrs.push_back(a);
+      }
+    }
+
+    auto extracted = ExtractBoundary2DSubmesh(*smesh, attr_list, internal_bdr_attrs, frame);
+
+    // Bake "other waveports act as PEC on this cross-section" into the mesh so the
+    // operator sees plain PEC edges and has no frame-dependent branch. Per-waveport
+    // frames compose this way without mutating iodata.
+    if (!other_waveport_attrs.empty())
+    {
+      MFEM_VERIFY(!bdr.pec.attributes.empty(),
+                  "BoundaryMode submesh extraction found other-waveport edges on the "
+                  "cross-section but no PEC attribute is defined in the config to "
+                  "relabel them to \u2014 define at least one PEC boundary attribute.");
+      const int pec_attr =
+          *std::min_element(bdr.pec.attributes.begin(), bdr.pec.attributes.end());
+      int relabelled = 0;
+      for (int sbe = 0; sbe < extracted->GetNBE(); sbe++)
+      {
+        if (other_waveport_attrs.count(extracted->GetBdrAttribute(sbe)))
+        {
+          extracted->SetBdrAttribute(sbe, pec_attr);
+          relabelled++;
+        }
+      }
+      if (relabelled > 0)
+      {
+        Mpi::Print(" Relabelled {:d} other-waveport boundary edge(s) as PEC (attr {:d})\n",
+                   relabelled, pec_attr);
+      }
+    }
+
+    smesh = std::move(extracted);
+    Mpi::Print(" Surface normal = ({:+.3e}, {:+.3e}, {:+.3e})\n", frame.normal(0),
+               frame.normal(1), frame.normal(2));
+  }
+
+  // Broadcast user-units frame to non-holding ranks.
+  Mpi::Broadcast(3, frame.centroid.HostReadWrite(), 0, comm);
+  Mpi::Broadcast(3, frame.e1.HostReadWrite(), 0, comm);
+  Mpi::Broadcast(3, frame.e2.HostReadWrite(), 0, comm);
+  Mpi::Broadcast(3, frame.normal.HostReadWrite(), 0, comm);
+
+  // Bake the tangent frame into iodata (pre-nondim, user units throughout):
+  //   - material tensors rotated into the local frame so MaterialOperator constructs
+  //     natively-2D downstream;
+  //   - impedance / voltage 3D path coordinates projected to 2D local coords so
+  //     PostOperator parses them as 2D paths directly.
+  // All subsequent downstream code is frame-agnostic.
+  RotateMaterialDefinitions(iodata.domains.materials, frame.e1, frame.e2, frame.normal);
+  for (auto &[idx, imp] : iodata.boundaries.postpro.impedance)
+  {
+    ProjectPathTo2D(imp.voltage_path, frame.centroid, frame.e1, frame.e2);
+    ProjectPathTo2D(imp.current_path, frame.centroid, frame.e1, frame.e2);
+  }
+  for (auto &[idx, vol] : iodata.boundaries.postpro.voltage)
+  {
+    ProjectPathTo2D(vol.voltage_path, frame.centroid, frame.e1, frame.e2);
+  }
+
+  // Resolve Lc from the 2D solve mesh and nondimensionalize in one pass.
+  BaseSolver::Preprocess(iodata, smesh, comm);
+}
+
 std::pair<ErrorIndicator, long long int>
 BoundaryModeSolver::Solve(const std::vector<std::unique_ptr<Mesh>> &mesh) const
 {
-  const auto &bm_data = iodata.solver.boundary_mode;
-  const int num_modes = bm_data.n;
-  const double freq_GHz = bm_data.freq;
+  const auto &bm = iodata.solver.boundary_mode;
+  const int num_modes = bm.n;
+  const double freq_GHz = bm.freq;
   const double omega =
       2.0 * M_PI * iodata.units.Nondimensionalize<Units::ValueType::FREQUENCY>(freq_GHz);
 
@@ -25,35 +214,53 @@ BoundaryModeSolver::Solve(const std::vector<std::unique_ptr<Mesh>> &mesh) const
              "(omega = {:.6e})\n",
              freq_GHz, omega);
 
-  // Construct the boundary mode operator (mesh, spaces, materials, eigen solver).
+  // MaterialOperator constructs natively-2D from (possibly pre-rotated) iodata; no
+  // post-hoc rotation is needed here.
   BlockTimer bt0(Timer::CONSTRUCT);
-  BoundaryModeOperator mode_op(iodata, mesh);
+  MaterialOperator mat_op(iodata, *mesh.back());
 
-  // Construct PostOperator.
-  PostOperator<ProblemType::BOUNDARYMODE> post_op(iodata, mode_op);
-  if (mode_op.IsFromSubmesh())
+  BoundaryModeOperator mode_op(iodata, mesh, mat_op);
+
+  // Combined block system DBC tdof list: ND indices, then H1 indices offset by nd_size.
+  const int nd_size = mode_op.GetNDTrueVSize();
+  mfem::Array<int> dbc_tdof_list;
+  dbc_tdof_list.Append(mode_op.GetNDDbcTDofLists().back());
+  for (int i = 0; i < mode_op.GetH1DbcTDofLists().back().Size(); i++)
   {
-    post_op.ProjectImpedancePaths(mode_op.GetSubmeshCentroid(), mode_op.GetSubmeshE1(),
-                                  mode_op.GetSubmeshE2());
-    post_op.ProjectVoltagePaths(mode_op.GetSubmeshCentroid(), mode_op.GetSubmeshE1(),
-                                mode_op.GetSubmeshE2());
+    dbc_tdof_list.Append(nd_size + mode_op.GetH1DbcTDofLists().back()[i]);
   }
 
-  // Error indicator (estimator is owned by mode_op).
-  auto &nd_fespace = mode_op.GetNDSpace();
+  if (const auto n_levels = mode_op.GetNDSpaceHierarchy().GetNumLevels(); n_levels > 1)
+  {
+    Mpi::Print(" Using p-multigrid preconditioning with {:d} levels\n",
+               static_cast<int>(n_levels));
+  }
+  const auto which_eig = (bm.target > 0.0) ? EigenvalueSolver::WhichType::LARGEST_MAGNITUDE
+                                           : EigenvalueSolver::WhichType::LARGEST_REAL;
+  ModeEigenSolver eig(mode_op, dbc_tdof_list, num_modes, bm.max_size, bm.tol, which_eig,
+                      iodata.solver.linear, bm.type, iodata.problem.verbose);
+
+  BoundaryModeFluxErrorEstimator<ComplexVector> estimator(
+      mode_op.GetMaterialOp(), mode_op.GetNDSpaceHierarchy(), mode_op.GetRTSpaceHierarchy(),
+      mode_op.GetCurlSpace(), mode_op.GetH1SpaceHierarchy(),
+      iodata.solver.linear.estimator_tol, iodata.solver.linear.estimator_max_it, 0,
+      iodata.solver.linear.estimator_mg);
+
+  PostOperator<ProblemType::BOUNDARYMODE> post_op(iodata, mode_op);
+
   ErrorIndicator indicator;
 
   // Determine kn_target.
   double kn_target;
-  if (bm_data.target > 0.0)
+  if (bm.target > 0.0)
   {
-    kn_target = bm_data.target * omega;
-    Mpi::Print(" Target n_eff = {:.6e}, kn_target = {:.6e}\n", bm_data.target, kn_target);
+    kn_target = bm.target * omega;
+    Mpi::Print(" Target n_eff = {:.6e}, kn_target = {:.6e}\n", bm.target, kn_target);
   }
   else
   {
     double c_min = mode_op.GetMaterialOp().GetLightSpeedMax().Min();
-    Mpi::GlobalMin(1, &c_min, nd_fespace.GetComm());
+    Mpi::GlobalMin(1, &c_min, mode_op.GetComm());
     MFEM_VERIFY(c_min > 0.0 && c_min < mfem::infinity(),
                 "Invalid material speed of light!");
     kn_target = omega / c_min * std::sqrt(1.1);
@@ -63,114 +270,72 @@ BoundaryModeSolver::Solve(const std::vector<std::unique_ptr<Mesh>> &mesh) const
   // Solve the eigenvalue problem.
   BlockTimer bt1(Timer::EPS);
   Mpi::Print("\nSolving GEP for {:d} propagation mode(s)...\n", num_modes);
-  auto result = mode_op.Solve(omega, kn_target);
+  const double sigma = -kn_target * kn_target;
+  auto result = eig.Solve(omega, sigma);
   int num_conv = result.num_converged;
-  double sigma = result.sigma;
   Mpi::Print(" Found {:d} converged eigenvalue{} (sigma = {:.6e})\n", num_conv,
-             (num_conv != 1) ? "s" : "", sigma);
+             (num_conv != 1) ? "s" : "", result.sigma);
 
   for (int i = 0; i < num_conv; i++)
   {
-    std::complex<double> lambda = mode_op.GetEigenvalue(i);
-    std::complex<double> kn = std::sqrt(-sigma - 1.0 / lambda);
+    auto kn = eig.GetPropagationConstant(i);
     Mpi::Print(" eig {:d}: kn = {:.6e}{:+.6e}i, n_eff = {:.6e}{:+.6e}i\n", i, kn.real(),
                kn.imag(), kn.real() / omega, kn.imag() / omega);
   }
 
   // Postprocessing.
   BlockTimer bt2(Timer::POSTPRO);
-  if (const auto *ksp = mode_op.GetLinearSolver())
+  if (const auto *ksp = eig.GetLinearSolver())
   {
     SaveMetadata(*ksp);
   }
   Mpi::Print("\nComputing solution error estimates and performing postprocessing\n\n");
 
-  const int nd_size = mode_op.GetNDTrueVSize();
   const int h1_size = mode_op.GetH1TrueVSize();
-  auto &l2_curl_fespace = mode_op.GetCurlSpace();
-  const auto &CurlOp = l2_curl_fespace.GetDiscreteInterpolator(nd_fespace);
-  const int l2_size = l2_curl_fespace.GetTrueVSize();
+  const int l2_size = mode_op.GetCurlSpace().GetTrueVSize();
 
   const int n_print = std::min(num_conv, num_modes);
   for (int i = 0; i < n_print; i++)
   {
-    std::complex<double> lambda = mode_op.GetEigenvalue(i);
-    std::complex<double> kn = std::sqrt(-sigma - 1.0 / lambda);
-    double error_bkwd = mode_op.GetError(i, EigenvalueSolver::ErrorType::BACKWARD);
-    double error_abs = mode_op.GetError(i, EigenvalueSolver::ErrorType::ABSOLUTE);
-
+    // Load eigenvector i, apply the shared VD back-transform so en holds physical En.
     ComplexVector e0(nd_size + h1_size);
     e0.UseDevice(true);
-    mode_op.GetEigenvector(i, e0);
+    eig.GetEigenvector(i, e0);
     ComplexVector et, en;
-    et.Real().MakeRef(e0.Real(), 0, nd_size);
-    et.Imag().MakeRef(e0.Imag(), 0, nd_size);
-    en.Real().MakeRef(e0.Real(), nd_size, h1_size);
-    en.Imag().MakeRef(e0.Imag(), nd_size, h1_size);
+    const std::complex<double> kn = eig.GetPropagationConstant(i);
+    mode_op.ApplyVDBackTransform(e0, kn, et, en);
 
-    // Power-normalize eigenvector using the full Poynting integral:
-    //   P = (1/2) Re{conj(kn)/ω × et^H Btt et} + Re{1/(2ωkn) × et^H Atn ẽn}
-    // The second term accounts for the Et·∇t(En) cross-coupling. At this point en = ẽn
-    // (VD variable, not yet back-transformed to physical En).
+    // Power-normalize the mode to |P| = 1. PostOperator re-evaluates P on the
+    // normalized mode for its impedance postprocessing.
+    const std::complex<double> P_initial = mode_op.ComputePoyntingPower(omega, kn, et, en);
+    if (std::abs(P_initial) > 0.0)
     {
-      const auto *Btt = mode_op.GetBtt();
-      if (Btt)
-      {
-        auto etH_Btt_et = linalg::Dot(nd_fespace.GetComm(), et, *Btt, et);
-        std::complex<double> P = 0.5 * std::conj(kn) / omega * etH_Btt_et;
-
-        // Cross-term: Re{1/(2ωkn) × et^H Atn ẽn}.
-        const auto *Atnr = mode_op.GetAtnr();
-        if (Atnr)
-        {
-          ComplexWrapperOperator Atn(Atnr, mode_op.GetAtni());
-          P += 1.0 / (2.0 * omega * kn) * linalg::Dot(nd_fespace.GetComm(), en, Atn, et);
-        }
-
-        double P_abs = std::abs(P);
-        if (P_abs > 0.0)
-        {
-          e0 *= 1.0 / std::sqrt(P_abs);
-        }
-        else
-        {
-          // Fallback to Btt mass norm (etH_Btt_et is real for symmetric Btt).
-          double norm2 = etH_Btt_et.real();
-          if (norm2 > 0.0)
-          {
-            e0 *= 1.0 / std::sqrt(norm2);
-          }
-        }
-      }
+      e0 *= 1.0 / std::sqrt(std::abs(P_initial));
     }
 
-    // Back-transform en from VD variable ẽn = ikn·En to physical En = ẽn/(ikn).
-    {
-      auto ikn_inv = 1.0 / (std::complex<double>(0.0, 1.0) * kn);
-      ComplexVector::AXPBY(ikn_inv, en.Real(), en.Imag(), 0.0, en.Real(), en.Imag());
-    }
+    const double error_bkwd = eig.GetError(i, EigenvalueSolver::ErrorType::BACKWARD);
+    const double error_abs = eig.GetError(i, EigenvalueSolver::ErrorType::ABSOLUTE);
 
     auto total_domain_energy =
         post_op.MeasureAndPrintAll(i, et, en, kn, omega, error_abs, error_bkwd, n_print);
 
-    const bool is_propagating =
-        std::abs(kn.imag()) < 0.1 * std::abs(kn.real()) && std::abs(kn.real()) > 0.0;
-    if (i < num_modes && is_propagating)
+    if (i < num_modes && ModeEigenSolver::IsPropagating(kn))
     {
+      // Bz = curl(Et) / (i·omega) = (Im(curl Et) - i·Re(curl Et)) / omega.
       ComplexVector bz(l2_size);
       bz.UseDevice(true);
-      {
-        Vector curl_etr(l2_size), curl_eti(l2_size);
-        curl_etr.UseDevice(true);
-        curl_eti.UseDevice(true);
-        CurlOp.Mult(et.Real(), curl_etr);
-        CurlOp.Mult(et.Imag(), curl_eti);
-        bz.Real() = curl_eti;
-        bz.Real() *= 1.0 / omega;
-        bz.Imag() = curl_etr;
-        bz.Imag() *= -1.0 / omega;
-      }
-      mode_op.AddErrorIndicator(et, bz, total_domain_energy, indicator);
+      const auto &CurlOp =
+          mode_op.GetCurlSpace().GetDiscreteInterpolator(mode_op.GetNDSpace());
+      Vector curl_etr(l2_size), curl_eti(l2_size);
+      curl_etr.UseDevice(true);
+      curl_eti.UseDevice(true);
+      CurlOp.Mult(et.Real(), curl_etr);
+      CurlOp.Mult(et.Imag(), curl_eti);
+      bz.Real() = curl_eti;
+      bz.Real() *= 1.0 / omega;
+      bz.Imag() = curl_etr;
+      bz.Imag() *= -1.0 / omega;
+      estimator.AddErrorIndicator(et, bz, total_domain_energy, indicator);
     }
   }
   Mpi::Print("\n");
