+++ palace/models/postoperator.cpp
@@ -1652,6 +1652,11 @@ auto PostOperator<solver_t>::MeasureAndPrintAll(int step, const ComplexVector &e
     -> std::enable_if_t<U == ProblemType::BOUNDARYMODE, double>
 {
   BlockTimer bt0(Timer::POSTPRO);
+
+  // Poynting power for impedance postprocessing. After the driver's power normalization
+  // |P| ≈ 1, but phase matters for the Z0 = V*conj(V) / (2P) formula.
+  const std::complex<double> P = fem_op->ComputePoyntingPower(omega, kn, et, en);
+
   SetEGridFunction(et);
 
   // Set the normal (out-of-plane) E component on H1 space.
@@ -1793,41 +1798,8 @@ auto PostOperator<solver_t>::MeasureAndPrintAll(int step, const ComplexVector &e
     return I;
   };
 
-  // Compute Poynting power P = (1/2) Re{kn*/ω} (et^H Btt et) — shared by all impedance
-  // entries.
-  std::complex<double> P(0.0, 0.0);
-  const auto *Btt = fem_op->GetBtt();
-  const auto *Atnr = fem_op->GetAtnr();
-  const auto *Atni = fem_op->GetAtni();
-  if (Btt != nullptr && !impedance_postpro.empty())
-  {
-    auto &nd_fespace = fem_op->GetNDSpace();
-    const int nd_size = nd_fespace.GetTrueVSize();
-    ComplexVector et_tdof(nd_size);
-    et_tdof.UseDevice(true);
-    E->Real().GetTrueDofs(et_tdof.Real());
-    E->Imag().GetTrueDofs(et_tdof.Imag());
-    P = 0.5 * std::conj(kn) / omega *
-        linalg::Dot(nd_fespace.GetComm(), et_tdof, *Btt, et_tdof);
-
-    // Cross-term: -Im{et^H Atn En} / (2ω). At this point En is physical (back-transformed).
-    if (Atnr != nullptr && En)
-    {
-      auto &h1_fespace = fem_op->GetH1Space();
-      const int h1_size = h1_fespace.GetTrueVSize();
-      ComplexVector en_tdof(h1_size);
-      en_tdof.UseDevice(true);
-      En->Real().GetTrueDofs(en_tdof.Real());
-      En->Imag().GetTrueDofs(en_tdof.Imag());
-      ComplexWrapperOperator Atn(Atnr, Atni);
-      auto etH_Atn_En = linalg::Dot(nd_fespace.GetComm(), en_tdof, Atn, et_tdof);
-      // For physical En: P_cross = (1/2) Re{-i/ω × conj(et^H Atn En)}
-      //                          = -Im{et^H Atn En} / (2ω)
-      P += std::complex<double>(-etH_Atn_En.imag(), etH_Atn_En.real()) / (2.0 * omega);
-    }
-  }
-
-  // Compute impedance for each configured entry.
+  // Compute impedance for each configured entry. P is the Poynting power of the
+  // power-normalized mode (|P| ≈ 1), computed by the caller on the mode eigensolver.
   for (const auto &[idx, cfg] : impedance_postpro)
   {
     auto V = ComputeVoltage(cfg.voltage_path, cfg.has_voltage_coordinates,
@@ -1894,47 +1866,6 @@ auto PostOperator<solver_t>::MeasureAndPrintAll(int step, const ComplexVector &e
     Mpi::Print(" Wrote mode {:d} to disk (grid function)\n", print_idx);
   }
   return measurement_cache.domain_E_field_energy_all;
-}
-
-template <ProblemType solver_t>
-void PostOperator<solver_t>::ProjectImpedancePaths(const mfem::Vector &centroid,
-                                                   const mfem::Vector &e1,
-                                                   const mfem::Vector &e2)
-{
-  for (auto &[idx, cfg] : impedance_postpro)
-  {
-    for (auto &p : cfg.voltage_path)
-    {
-      if (p.Size() == 3)
-      {
-        p = mesh::Project3Dto2D(p, centroid, e1, e2);
-      }
-    }
-    for (auto &p : cfg.current_path)
-    {
-      if (p.Size() == 3)
-      {
-        p = mesh::Project3Dto2D(p, centroid, e1, e2);
-      }
-    }
-  }
-}
-
-template <ProblemType solver_t>
-void PostOperator<solver_t>::ProjectVoltagePaths(const mfem::Vector &centroid,
-                                                 const mfem::Vector &e1,
-                                                 const mfem::Vector &e2)
-{
-  for (auto &[idx, cfg] : voltage_postpro)
-  {
-    for (auto &p : cfg.voltage_path)
-    {
-      if (p.Size() == 3)
-      {
-        p = mesh::Project3Dto2D(p, centroid, e1, e2);
-      }
-    }
-  }
 }
 
 // Explicit template instantiation.
