+++ palace/models/waveportoperator.hpp
@@ -16,6 +16,7 @@
 #include "linalg/ksp.hpp"
 #include "linalg/operator.hpp"
 #include "linalg/vector.hpp"
+#include "models/modeeigensolver.hpp"
 #include "utils/configfile.hpp"
 #include "utils/labels.hpp"
 
@@ -44,230 +45,7 @@ struct WavePortData;
 
 enum class ProblemType : char;
 
-//
-// Helper class for wave port boundaries in a model.
-//
-template <typename OperType>
-class BlockDiagonalPreconditioner;
 class FiniteElementSpace;
-class FiniteElementSpaceHierarchy;
-
-// (ModeEigenSolverConfig removed — solver parameters passed directly to constructor.)
-
-//
-// Configuration for p-multigrid preconditioning of the boundary mode eigenvalue problem.
-// When provided, the ModeEigenSolver uses a block-diagonal GMG preconditioner
-// (ND p-multigrid + H1 p-multigrid) instead of a sparse direct solver.
-//
-struct ModeEigenSolverMultigridConfig
-{
-  // ND and H1 FE space hierarchies for the diagonal blocks (not owned).
-  FiniteElementSpaceHierarchy *nd_fespaces = nullptr;
-  FiniteElementSpaceHierarchy *h1_fespaces = nullptr;
-
-  // H1 auxiliary space hierarchy for Hiptmair distributive relaxation in the ND block.
-  FiniteElementSpaceHierarchy *h1_aux_fespaces = nullptr;
-
-  // Per-level essential BC true DOF lists for ND, H1, and H1 auxiliary blocks (not owned).
-  std::vector<mfem::Array<int>> *nd_dbc_tdof_lists = nullptr;
-  std::vector<mfem::Array<int>> *h1_dbc_tdof_lists = nullptr;
-  std::vector<mfem::Array<int>> *h1_aux_dbc_tdof_lists = nullptr;
-};
-
-//
-// Linear eigenvalue solver for 2D boundary mode computation using a direct linearization
-// of the transverse curl-curl (Equation 1) and normal curl-curl (Equation 2) equations
-// with the Vardapetyan-Demkowicz substitution e_n_tilde = i*kn*E_n, e_t = E_t.
-//
-// Assembles and solves the generalized eigenvalue problem:
-//
-//   [Att  Atn] [et]           [Btt  Btn] [et]
-//   [ 0   Ann] [en] = lambda  [ 0    0 ] [en]
-//
-// where:
-//   Att = mu_cc^{-1}(curl_t u, curl_t v) - omega^2(eps u, v) + BC-t
-//   Atn = -(mu^{-1} grad_t u, v)
-//   Ann = -(mu^{-1} grad u, grad v) + omega^2(eps u, v) + BC-n (H1 stiffness + mass +
-//          impedance boundary)
-//   Btt = (mu^{-1} u, v)
-//   Btn = -Atn^T (negative transpose coupling from Equation 2)
-//
-// The A matrix is upper block-triangular (Ant = 0). The B matrix has Btn coupling from
-// Equation 2 and Bnn = 0. This is a standard GEP with shift-and-invert.
-//
-// The eigenvector contains [e_t; e_n_tilde] where e_n_tilde = i*kn*E_n. Recovery of
-// E_n = e_n_tilde / (i*kn) is performed in the driver, not here.
-//
-class ModeEigenSolver
-{
-public:
-  struct SolveResult
-  {
-    int num_converged;
-    double sigma;
-  };
-
-  // The constructor assembles frequency-independent matrices (Atn, Btn = Atn^T, Btt)
-  // and configures linear and eigenvalue solvers. Frequency-dependent matrices (Att, Ann)
-  // are assembled at solve time via AssembleFrequencyDependent(). Matrix assembly uses
-  // the FE space communicator (all processes). If solver_comm != MPI_COMM_NULL, the
-  // linear and eigenvalue solvers are configured on that communicator.
-  // Constructor. Material properties and attribute mapping come from mat_op directly.
-  // For 3D wave port submeshes, normal is the outward surface normal; for 2D domain
-  // meshes, normal is nullptr. Boundary operators are optional (nullptr = none).
-  ModeEigenSolver(const MaterialOperator &mat_op, const mfem::Vector *normal,
-                  SurfaceImpedanceOperator *surf_z_op,
-                  FarfieldBoundaryOperator *farfield_op,
-                  SurfaceConductivityOperator *surf_sigma_op,
-                  const FiniteElementSpace &nd_fespace,
-                  const FiniteElementSpace &h1_fespace,
-                  const mfem::Array<int> &dbc_tdof_list, int num_modes, int num_vec,
-                  double eig_tol, EigenvalueSolver::WhichType which_eig,
-                  const config::LinearSolverData &linear, EigenSolverBackend eigen_backend,
-                  int verbose, MPI_Comm solver_comm = MPI_COMM_NULL,
-                  const ModeEigenSolverMultigridConfig *mg_config = nullptr);
-
-  ~ModeEigenSolver();
-
-  // Assemble frequency-dependent matrices and solve the eigenvalue problem. The shift
-  // sigma = -kn_target^2 is applied. An optional initial space vector can be provided
-  // for eigenvalue solver warm-starting. When has_solver is true (default), the calling
-  // process participates in the eigenvalue solve; when false (wave port non-port process),
-  // only assembly is performed.
-  SolveResult Solve(double omega, double sigma, bool has_solver = true,
-                    const ComplexVector *initial_space = nullptr);
-
-  // Access converged eigenvalues and eigenvectors.
-  std::complex<double> GetEigenvalue(int i) const;
-  void GetEigenvector(int i, ComplexVector &x) const;
-
-  // Get the eigenpair error for the i-th converged eigenvalue.
-  double GetError(int i, EigenvalueSolver::ErrorType type) const;
-
-  // Access the assembled Btt and Atn matrices (needed for power normalization).
-  const mfem::HypreParMatrix *GetBtt() const { return Bttr.get(); }
-  const mfem::HypreParMatrix *GetAtnr() const { return Atnr.get(); }
-  const mfem::HypreParMatrix *GetAtni() const { return Atni.get(); }
-
-  // Get the true vector sizes for the ND and H1 FE spaces.
-  int GetNDTrueVSize() const { return nd_size; }
-  int GetH1TrueVSize() const { return h1_size; }
-
-  // Access the linear solver (for metadata reporting). Returns nullptr if this process
-  // does not have a solver configured (non-port process in wave port mode).
-  const ComplexKspSolver *GetLinearSolver() const { return ksp.get(); }
-
-private:
-  // Solver parameters.
-  int num_modes, num_vec;
-  double eig_tol;
-  EigenvalueSolver::WhichType which_eig;
-  const config::LinearSolverData &linear;
-  EigenSolverBackend eigen_backend;
-  int verbose;
-
-  // Material operator and boundary operators (not owned).
-  const MaterialOperator &mat_op;
-  const mfem::Vector *normal;
-  SurfaceImpedanceOperator *surf_z_op;
-  FarfieldBoundaryOperator *farfield_op;
-  SurfaceConductivityOperator *surf_sigma_op;
-
-  // References to FE spaces (not owned).
-  const FiniteElementSpace &nd_fespace;
-  const FiniteElementSpace &h1_fespace;
-
-  // Essential boundary condition true DOF list for the combined block system.
-  mfem::Array<int> dbc_tdof_list;
-
-  // Cached FE space sizes.
-  int nd_size, h1_size;
-
-  // Frequency-independent assembled matrices.
-  // Atn: gradient coupling -(mu^{-1} grad_t u, v).
-  // Btn: negative transpose of Atn: (mu^{-1} u, grad_t v).
-  // Bttr: positive copy of Btt for external use (postprocessing).
-  std::unique_ptr<mfem::HypreParMatrix> Atnr, Atni;
-  std::unique_ptr<mfem::HypreParMatrix> Btnr, Btni;
-  std::unique_ptr<mfem::HypreParMatrix> Bttr;
-  std::unique_ptr<ComplexOperator> opB;
-
-  // Frequency-dependent block A operator (rebuilt each Solve).
-  std::unique_ptr<ComplexOperator> opA;
-
-  // Eigenvalue and linear solvers (null on processes without solver_comm).
-  std::unique_ptr<EigenvalueSolver> eigen;
-  std::unique_ptr<ComplexKspSolver> ksp;
-
-  // Permutation that maps external mode index to eigensolver index, sorted by ascending
-  // Re{kn}. This ensures consistent mode ordering regardless of eigensolver backend.
-  std::vector<int> mode_perm;
-
-  // Assemble frequency-dependent Att and Ann, then build block A (MPI collective on FE
-  // space comm).
-  void AssembleFrequencyDependent(double omega, double sigma);
-
-  // Private helper methods for bilinear form assembly.
-  using ComplexHypreParMatrix = std::tuple<std::unique_ptr<mfem::HypreParMatrix>,
-                                           std::unique_ptr<mfem::HypreParMatrix>>;
-
-  // Att: curl-curl + mass + BC-t.
-  ComplexHypreParMatrix AssembleAtt(double omega, double sigma) const;
-
-  // Atn: gradient coupling -(mu^{-1} grad_t u, v).
-  ComplexHypreParMatrix AssembleAtn() const;
-
-  // Ann: H1 stiffness + mass + BC-n.
-  // Ann = -(mu^{-1} grad u, grad v) + omega^2(eps u, v) + BC-n impedance.
-  ComplexHypreParMatrix AssembleAnn(double omega) const;
-
-  // Btt: ND mass (mu^{-1} u, v) (positive).
-  ComplexHypreParMatrix AssembleBtt() const;
-
-  // Build the 2x2 block A matrix. The (1,0) block is -sigma * Btn from the
-  // shift-and-invert transformation (nullptr when sigma = 0).
-  ComplexHypreParMatrix
-  BuildSystemMatrixA(const mfem::HypreParMatrix *Attr, const mfem::HypreParMatrix *Atti,
-                     const mfem::HypreParMatrix *Atnr, const mfem::HypreParMatrix *Atni,
-                     const mfem::HypreParMatrix *Annr, const mfem::HypreParMatrix *Anni,
-                     const mfem::HypreParMatrix *shifted_Btnr = nullptr) const;
-
-  // Build the 2x2 block B matrix: [Btt, 0; Btn, 0].
-  ComplexHypreParMatrix BuildSystemMatrixB(const mfem::HypreParMatrix *Bttr,
-                                           const mfem::HypreParMatrix *Btti,
-                                           const mfem::HypreParMatrix *Btnr,
-                                           const mfem::HypreParMatrix *Btni,
-                                           const mfem::HypreParMatrix *Dnn) const;
-
-  // Optional multigrid configuration (not owned, may be nullptr for sparse direct path).
-  const ModeEigenSolverMultigridConfig *mg_config;
-
-  // Non-owning pointer to the block preconditioner (for setting operators in Solve).
-  BlockDiagonalPreconditioner<ComplexOperator> *block_pc_ptr = nullptr;
-
-  // Multigrid preconditioner operators (owned, must outlive the GMG solver application).
-  std::unique_ptr<ComplexMultigridOperator> att_mg_op, ann_mg_op;
-
-  // Shifted off-diagonal operator -sigma*Btn for block lower-triangular preconditioning.
-  std::unique_ptr<ComplexOperator> shifted_Btn_op;
-
-  // Assemble preconditioner operators at all multigrid levels for the Att (ND) block.
-  // Returns a ComplexMultigridOperator with primary (ND) and auxiliary (H1) operators.
-  std::unique_ptr<ComplexMultigridOperator> AssembleAttPreconditioner(double omega,
-                                                                      double sigma) const;
-
-  // Assemble preconditioner operators at all multigrid levels for the Ann (H1) block.
-  std::unique_ptr<ComplexMultigridOperator> AssembleAnnPreconditioner(double omega) const;
-
-  // Set up the linear solver (GMRES + sparse direct preconditioner).
-  void SetUpLinearSolver(MPI_Comm comm);
-
-  // Set up the linear solver with p-multigrid block-diagonal preconditioning.
-  void SetUpMultigridLinearSolver(MPI_Comm comm);
-
-  // Set up the eigenvalue solver (SLEPc or ARPACK).
-  void SetUpEigenSolver(MPI_Comm comm);
-};
 
 //
 // Wave port data for a single port boundary.
@@ -324,11 +102,8 @@ private:
 
   // Voltage path for line integral (optional, for impedance postprocessing).
   std::vector<mfem::Vector> voltage_path;
-  int voltage_n_samples = 100;
+  int voltage_n_samples;
   bool has_voltage_coords = false;
-
-  // Whether this is a 2D direct port (no submesh).
-  bool is_2d_direct = false;
 
 public:
   // 3D submesh constructor: extracts submesh from parent mesh.
@@ -338,18 +113,6 @@ public:
                const MaterialOperator &mat_op, mfem::ParFiniteElementSpace &nd_fespace,
                mfem::ParFiniteElementSpace &h1_fespace, const mfem::Array<int> &dbc_attr);
 
-  // 2D direct constructor: uses provided mesh/spaces directly. No submesh extraction,
-  // no parent DOF transfer. Used by BoundaryModeOperator.
-  WavePortData(const MaterialOperator &mat_op, const mfem::Vector *normal,
-               SurfaceImpedanceOperator *surf_z_op, FarfieldBoundaryOperator *farfield_op,
-               SurfaceConductivityOperator *surf_sigma_op, FiniteElementSpace &nd_fespace,
-               FiniteElementSpace &h1_fespace, const mfem::Array<int> &dbc_tdof_list,
-               int num_modes, int num_vec, double eig_tol,
-               EigenvalueSolver::WhichType which_eig,
-               const config::LinearSolverData &linear, EigenSolverBackend eigen_backend,
-               int verbose, MPI_Comm solver_comm = MPI_COMM_NULL,
-               const ModeEigenSolverMultigridConfig *mg_config = nullptr);
-
   ~WavePortData();
 
   [[nodiscard]] constexpr bool HasExcitation() const { return excitation != 0; }
@@ -358,11 +121,6 @@ public:
   const auto &GetAttrList() const { return attr_list; }
 
   void Initialize(double omega);
-
-  // Access the eigenvalue solver directly (for 2D direct mode used by
-  // BoundaryModeOperator).
-  ModeEigenSolver &GetEigenSolver() { return *mode_solver; }
-  const ModeEigenSolver &GetEigenSolver() const { return *mode_solver; }
 
   HYPRE_BigInt GlobalTrueNDSize() const { return port_nd_fespace->GlobalTrueVSize(); }
   HYPRE_BigInt GlobalTrueH1Size() const { return port_h1_fespace->GlobalTrueVSize(); }
